% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBWidget-class.R, R/src_sound.R
\docType{class}
\name{DBWidget-class}
\alias{DBWidget-class}
\alias{.DBWidget}
\alias{DBWidget}
\alias{sql_file}
\alias{src_tbls.src_sound}
\alias{tbl.src_sound}
\alias{tbl.DBWidget}
\alias{show,DBWidget-method}
\alias{sql_file}
\alias{src_tbls.src_sound}
\alias{tbl.src_sound}
\title{Create and use a data base (sqlite)-backed widget}
\usage{
DBWidget(db_directory, sql_cmd_template)

sql_file(object)

\method{src_tbls}{src_sound}(x)

\method{tbl}{src_sound}(src, from, ...)

\method{tbl}{DBWidget}(src, from, ...)

\S4method{show}{DBWidget}(object)

sql_file(object)

\method{src_tbls}{src_sound}(x)

\method{tbl}{src_sound}(src, from, ...)
}
\arguments{
\item{sql_cmd_template}{character(1) path to existing SQL template
file, containing table and templated query definitions.}

\item{object}{An instance of class \code{DBWidget}}

\item{x}{An instance of class \code{DBWidget}.}

\item{src}{A \code{DBWidget} or \code{src_sound} instance.}

\item{...}{key-value pairs of case information. Arguments must be
named and unique. Each value must be a scalar. When
\code{length(allowed_keys) != 0}, names of keys must be in
\code{allowed_keys}.}

\item{allowed_keys}{character() of allowed key names.}

\item{sql_file}{character(1) path to SQL file. If the file does not
exist, the template from \code{sql_cmd_template} is used to
create the data. If the data base exists, no checks are
performed to validate atains the provide
\code{sql_cmd_template}.}

\item{object}{An instance of class \code{DBWidget}}

\item{x}{An instance of class \code{DBWidget}.}

\item{src}{A \code{DBWidget} or \code{src_sound} instance.}
}
\value{
\code{sql_file()} returns a dplyr 'src' object referencing
    the data base.

\code{src_tbls()} returns the names of user-accessible
    tables available in \code{x}.

\code{tbl()} returns the dplyr sqlite table corresponding
    to \code{from}.

\code{sql_file()} returns a dplyr 'src' object referencing
    the data base.

\code{src_tbls()} returns the names of user-accessible
    tables available in \code{x}.

\code{tbl()} returns the dplyr sqlite table corresponding
    to \code{from}.
}
\description{
This page describes a widget for working with data
    base backed resources. A typical use case is to manage a
    collection of board, case and assay results.

\code{DBWidget()}: create an instance of a data base widget.
}
