% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbsave-load-report.R
\docType{methods}
\name{sbsave}
\alias{sbsave}
\alias{sbload}
\alias{sbreport}
\alias{sbreport,tbl_board-method}
\alias{sbreport,tbl_cases-method}
\alias{sbreport,tbl_assay-method}
\alias{sbreport,data.frame-method}
\alias{sbreport,gg-method}
\title{Save, load, and report on SOUNDWidget instances.}
\usage{
sbsave(x, file)

sbload(x, file)

sbreport(x)

\S4method{sbreport}{tbl_board}(x)

\S4method{sbreport}{tbl_cases}(x)

\S4method{sbreport}{tbl_assay}(x)

\S4method{sbreport}{data.frame}(x)

\S4method{sbreport}{gg}(x)
}
\arguments{
\item{x}{An instance derived from \code{SOUNDWidget-class}.}

\item{file}{character(1) Location to or from which the instance
\code{x} is to be saved / loaded.}
}
\description{
These methods are usually created for the user with
    \code{SOUNDWidget()}, they are not created directly. Likewise
    methods are invoked by SOUNDBoard infrastructure, not by the
    user.
}
\details{
An \code{sbsave(x, file)} method saves \code{x} to \code{file}.

An \code{sbload(x, file)} method creates an instance of class
\code{x} from location \code{file}. The instance \code{x} passed to
the mehod is created by the default constructor of \code{x}, and
does \emph{not} contain the resource. A typical method will input
the data \code{resource <- ...}, and then populate the instance
\code{initialize(x, resource = resource)}

An \code{sbreport(x)} method creates output from \code{x} suitable
for inclusion in a SOUNDBoard report.
}
