% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOUNDBoardOption-class.R
\name{SOUNDBoardOption}
\alias{SOUNDBoardOption}
\alias{getSOUNDBoardOption}
\title{Manage SOUNDBoard options}
\usage{
SOUNDBoardOption(...)

getSOUNDBoardOption(option)
}
\arguments{
\item{...}{key-value pairs of valid SOUNDBoard options.
    \itemize{

        \item{server_host} character(1) Deployment SOUNDBoard
        server host name.

        \item{server_path} character(1) Full deployment path on
        server.

        \item{username} character(1) User name of SOUNDBoard
        deployment account. The user of the account performing the
        deployment must have passwordless login credentials to
        \code{username@server_host:server_path}.

    }}

\item{option}{character(1) Name of option to retrieve.}
}
\value{
\code{SOUNDBoardOption()} returns NULL, invisibly.

\code{getSOUNDBoardOption()} returns the character(1) value
    of the requested option.
}
\description{
This page describes options available in the
    SOUNDBoard package. Options can be set with named key-value
    pairs, e.g., \code{SOUNDBoardOption(username =
    "soundboard")}. Options can also be specified by environment
    variables formed by the contatentation of \code{SOUNDBOARD_}
    and the option name, e.g., in bash \code{export
    SOUNDBOARD_USERNAME=soundboard}.
}
