% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyWidget-class.R
\name{SOUNDBoardWidget}
\alias{SOUNDBoardWidget}
\title{Interactive display of SOUND resources through a the
ShinyAppWidget class}
\usage{
SOUNDBoardWidget(soundmgr)
}
\arguments{
\item{soundmgr}{A \linkS4class{SOUNDManager} class}
}
\description{
This function will display a \link{SOUNDManager-class} object
as an interactive shiny widget. The widget allows the user to select
by board, case and assay. The displays will react depending on selected
components of the database. The database is to be created beforehand.
}
