% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOUNDWidget-class.R
\docType{class}
\name{SOUNDWidget-class}
\alias{SOUNDWidget-class}
\alias{.SOUNDWidget}
\alias{sbresource}
\alias{SOUNDWidget}
\title{Create and use 'widgets' for consistent data presenation}
\usage{
sbresource(x)

SOUNDWidget(widget, save, load, report, ..., where = .GlobalEnv)
}
\arguments{
\item{x}{A \code{SOUNDWidget} instance.}

\item{widget}{character(1) widget class name}

\item{save}{function for saving the widget to disk}

\item{load}{function for loading the widget from disk}

\item{report}{method used to display widget after loading (i.e., 'sbreport')}

\item{...}{key-value pairs of case information. Arguments must be
named and unique. Each value must be a scalar. When
\code{length(allowed_keys) != 0}, names of keys must be in
\code{allowed_keys}.}

\item{where}{The environment in which to define a
SOUNDWidget-derived class and methods.}

\item{allowed_keys}{character() of allowed key names.}
}
\description{
This page describes 'widgets' for representating and
    reporting elements of a SOUND Board Report. Each widget
    describes a particular component of the data, e.g., the
    'metadata' about the case; a top table of differentially
    expressed genes, an interactive plot of variant effects. The
    widget provides a constructor method, as well as a 'render'
    method. Widgets themselves may perform considerable work when
    created, but render quickly. Widgets can be fully serialized.
}
