% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SOUNDManager-class.R
\docType{class}
\name{SOUNDManager-class}
\alias{SOUNDManager-class}
\alias{.SOUNDManager}
\alias{show,SOUNDBoard-method}
\alias{SOUNDManager}
\alias{show,SOUNDManager-method}
\title{Managing SOUNDBoard Reports}
\usage{
SOUNDManager(board_directory, sql_template_path, host = "localhost",
  port = "3838", path = "SOUNDBoard", username = "soundboard")

\S4method{show}{SOUNDManager}(object)
}
\arguments{
\item{board_directory}{character(1) Local directory where board, case,
and assay elements are marshalled.}

\item{sql_template_path}{character(1) path to existing SQL template
file, containing table and templated query definitions.}

\item{host}{character(1) URI of deployment server directory.}

\item{port}{character(1) communation endpoint from the host.}

\item{path}{character(1) path to directory where SOUNDBoard
reports are served.}

\item{username}{character(1) name of SOUNDBoard user from which
applications are deployed. Default: "soundboard"}

\item{object}{A \code{SOUNDManager} instance.}
}
\description{
This page describes facilities for managing SOUNDBoard
    reports. Management tasks include (1) creating template
    reports; (2) publishing as temporary documents during
    development; and (3) publishing as final reports for
    communciation with board members.
}
